+++ src/datastore.py
@@ -1,59 +1,31 @@
-import json
 import traceback
 
 import logging
 log = logging.getLogger(__name__)
 
-import couchbase
-from couchbase.exceptions import NotFoundError
+from couchbase.bucket import Bucket
+from couchbase.exceptions import NotFoundError, KeyExistsError
 
 
-class ClientBase(object):
-    shortcut = None
+class Datastore(object):
 
-    def __init__(self, host, port, database, databasePassword):
-        pass
-
-    def create(self):
-        pass
-
-    def read(self):
-        pass
-
-    def update(self):
-        pass
-
-    def delete(self):
-        pass
-
-
-class ClientCouchbase(ClientBase):
-    shortcut = "couchbase"
-
-    def __init__(self, host, port, database, databasePassword,
-                 jsonEncoder=None):
-
-        if jsonEncoder:
-            couchbase.set_json_converters(jsonEncoder, json.loads)
-
-        self.connection = couchbase.Couchbase.connect(
-            host=host,
-            port=port,
-            bucket=database,
-            password=databasePassword
-        )
+    def __init__(self, host, port, bucket, **kwargs):
+        connectionString = "http://%s:%s/%s" % (host, port, bucket)
+        self.bucket = Bucket(connectionString, **kwargs)
 
     def create(self, key, value):
         try:
-            self.connection.add(key, value)
+            self.bucket.insert(key, value)
             return True
+        except KeyExistsError:
+            return False
         except:
             log.error("unexpected error, %s" % traceback.format_exc())
             return False
 
     def read(self, key):
         try:
-            doc = self.connection.get(key)
+            doc = self.bucket.get(key)
             return doc.value
         except NotFoundError:
             return False
@@ -63,7 +35,7 @@ class ClientCouchbase(ClientBase):
 
     def update(self, key, value):
         try:
-            self.connection.replace(key, value)
+            self.bucket.replace(key, value)
             return True
         except NotFoundError:
             return False
@@ -73,7 +45,7 @@ class ClientCouchbase(ClientBase):
 
     def set(self, key, value):
         try:
-            self.connection.set(key, value)
+            self.bucket.upsert(key, value)
             return True
         except:
             log.error("unexpected error, %s" % traceback.format_exc())
@@ -81,7 +53,7 @@ class ClientCouchbase(ClientBase):
 
     def delete(self, key):
         try:
-            self.connection.delete(key)
+            self.bucket.remove(key)
             return True
         except NotFoundError:
             return False
@@ -90,65 +62,13 @@ class ClientCouchbase(ClientBase):
             return False
 
     def view(self, design, view, **kwargs):
-        return self.connection.query(design, view, **kwargs)
+        return self.bucket.query(design, view, **kwargs)
 
+    def get_multi(self, keys):
+        return self.bucket.get_multi(keys)
 
-class Datastore(object):
-    client = None
+    def design_get(self, name, **kwargs):
+        return self.bucket.bucket_manager().design_get(name, **kwargs)
 
-    def __init__(self, clientType, database=None, databasePassword=None,
-                 host="localhost", port=8091, **kwargs):
-
-        clientTypes = ClientBase.__subclasses__()
-        for ct in clientTypes:
-            # two checks: one for clientType as a class (ex: ClientCouchbase),
-            # one for clientType as a string ("couchbase")
-            if (
-                ct == clientType or
-                ct.__name__ == clientType or
-                ct.shortcut == clientType
-            ):
-                self.client = ct(
-                    host=host,
-                    port=port,
-                    database=database,
-                    databasePassword=databasePassword,
-                    **kwargs
-                )
-
-        if self.client is None:
-            raise ClientTypeNotFound(clientType)
-
-    def create(self, key, value):
-        return self.client.create(key, value)
-
-    def read(self, key):
-        return self.client.read(key)
-
-    def update(self, key, value):
-        return self.client.update(key, value)
-
-    def set(self, key, value):
-        return self.client.set(key, value)
-
-    def delete(self, key):
-        return self.client.delete(key)
-
-    def view(self, design, view, **kwargs):
-        return self.client.view(design, view, **kwargs)
-
-
-class ClientTypeNotFound(Exception):
-    def __init__(self, clientType):
-        clientsArray = ClientBase.__subclasses__()
-        clientsString = ""
-        for c in clientsArray:
-            clientsString += "\n%s" % (str(c.__name__))
-
-        self.message = "\nCould not find the client type: %s. " \
-                       "\nIf you need it, here is a list of the " \
-                       "available client types: %s\n" % \
-                       (clientType, clientsString)
-
-    def __str__(self):
-        return self.message
+    def design_create(self, name, ddoc, **kwargs):
+        return self.bucket.bucket_manager().design_create(name, ddoc, **kwargs)